+++ kernel/include/fs/buffer_manager.hh
@@ -47,13 +47,13 @@ namespace fs
 		BufferManager() {};
 		void init( const char *lock_name );
 
-		Buffer read( int dev, uint lba );
+		Buffer read( int dev, uint64 lba );
 
-		Buffer read_sync( int dev, uint lba );
+		Buffer read_sync( int dev, uint64 lba );
 
-		void write( int dev, uint lba );
+		void write( int dev, uint64 lba );
 
-		void write_sync( int dev, uint lba );
+		void write_sync( int dev, uint64 lba );
 
 		/// @brief 同步方法，比较费时，应当在调度开始前供内核使用
 		void release_buffer_sync( Buffer &buf )
